# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "灵嘚中间件（LingDei-Middleware）是一个基于 Fiber 的 RESTful API 服务，用于提供灵嘚（LingDei）的后端服务。\n注意，有 🦸 标识的接口需要管理员权限才能访问。",
        "title": "灵嘚中间件",
        "contact": {},
        "version": "v1.0.0"
    },
    "host": "127.0.0.1:9000",
    "basePath": "/",
    "paths": {
        "/barrage/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加Barrage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "弹幕管理"
                ],
                "summary": "添加Barrage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "弹幕内容",
                        "name": "content",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "弹幕秒数",
                        "name": "second",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/barrage/count": {
            "get": {
                "description": "获取Barrage数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "弹幕管理"
                ],
                "summary": "获取Barrage数量",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BarrageCountResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/barrage/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除Barrage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "弹幕管理"
                ],
                "summary": "删除Barrage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "弹幕UUID",
                        "name": "uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/barrage/get": {
            "get": {
                "description": "获取Barrage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "弹幕管理"
                ],
                "summary": "获取Barrage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "弹幕UUID",
                        "name": "uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BarrageResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/barrage/list": {
            "get": {
                "description": "获取Barrage列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "弹幕管理"
                ],
                "summary": "获取Barrage列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BarrageListResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/barrage/recent_list": {
            "get": {
                "description": "获取最近的Barrage列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "弹幕管理"
                ],
                "summary": "获取最近的Barrage列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "弹幕秒数",
                        "name": "second",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "几秒内的",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BarrageListResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
...
        "/collect/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "收藏视频",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "收藏管理"
                ],
                "summary": "收藏视频",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/collect/check": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "检查是否收藏过",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "收藏管理"
                ],
                "summary": "检查是否收藏过",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CollectStatusResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/collect/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "取消收藏",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "收藏管理"
                ],
                "summary": "取消收藏",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频UUID",
                        "name": "video_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/collect/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取某位用户自己的收藏列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "收藏管理"
                ],
                "summary": "获取某位用户自己的收藏列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CollectListResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Barrage": {
            "type": "object",
            "required": [
                "content",
                "second",
                "timestamp",
                "user_uuid",
                "uuid",
                "video_uuid"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "second": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "integer"
                },
                "user_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "video_uuid": {
                    "type": "string"
                }
            }
        },
        "model.BarrageCountResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "model.BarrageListResp": {
            "type": "object",
            "properties": {
                "barrage_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Barrage"
                    }
                },
                "code": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.BarrageResp": {
            "type": "object",
            "properties": {
                "barrage": {
                    "$ref": "#/definitions/model.Barrage"
                },
                "code": {
                    "type": "integer"
                }
            }
        },
...
        "model.Collect": {
            "type": "object",
            "required": [
                "user_uuid",
                "uuid",
                "video_uuid"
            ],
            "properties": {
                "user_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "video": {
                    "$ref": "#/definitions/model.Video"
                },
                "video_uuid": {
                    "type": "string"
                }
            }
        },
        "model.CollectListResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "collect_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Collect"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.CollectStatusResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "model.OperationResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.Video": {
            "type": "object",
            "required": [
                "author_uuid",
                "category_uuid",
                "name",
                "thumbnail_url",
                "timestamp",
                "url",
                "uuid"
            ],
            "properties": {
                "author_uuid": {
                    "type": "string"
                },
                "category_uuid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}